+++ type/PyBool.cs
@@ -302,7 +302,7 @@ namespace SharpPy
         /// <summary>
         /// CPython PyFloat_AsDouble 호환: PyBool에서 C# double 값 추출  
         /// </summary>
-        public override double ToFloat() => Value ? 1.0 : 0.0;
+        public override float ToFloat() => Value ? 1.0f : 0.0f;
         public override double ToDouble() => Value ? 1.0 : 0.0;
 
         /// <summary>
